% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_data.R
\name{generate_dummy_data_time}
\alias{generate_dummy_data_time}
\title{Title}
\usage{
generate_dummy_data_time(
  n_cells,
  sd_celltypes,
  n_major_cell_types,
  n_minor_cell_types,
  relative_abundance,
  n_major_interact_celltypes,
  n_minor_interact_celltypes,
  n_individuals,
  n_batchs,
  interaction_feature,
  time_points,
  test_var,
  prop_disease,
  fc_interact,
  interaction_type = c("specific", "differential", "opposite"),
  seed = 1234,
  visit_effects_progressor,
  visit_effects_control,
  direction_by_cluster
)
}
\arguments{
\item{n_cells}{}

\item{sd_celltypes}{}

\item{n_major_cell_types}{}

\item{n_minor_cell_types}{}

\item{relative_abundance}{}

\item{n_major_interact_celltypes}{}

\item{n_minor_interact_celltypes}{}

\item{n_individuals}{}

\item{n_batchs}{}

\item{interaction_feature}{}

\item{time_points}{}

\item{test_var}{}

\item{prop_disease}{}

\item{fc_interact}{}

\item{interaction_type}{}

\item{seed}{}

\item{visit_effects_progressor}{}

\item{visit_effects_control}{}

\item{direction_by_cluster}{}
}
\description{
Title
}
