% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_data_time_3plus.R
\name{generate_dummy_data_time_3plus}
\alias{generate_dummy_data_time_3plus}
\title{Title}
\usage{
generate_dummy_data_time_3plus(
  n_cells = 3000,
  sd_celltypes = 0.1,
  n_major_cell_types = 7,
  n_minor_cell_types = 3,
  relative_abundance = 0.1,
  n_major_interact_celltypes = 1,
  n_minor_interact_celltypes = 1,
  n_individuals = 30,
  n_batchs = 4,
  interaction_feature = c("visit"),
  time_points = 2,
  test_var = c("disease"),
  prop_disease = 0.5,
  fc_interact = 0.1,
  interaction_type = c("specific", "differential", "opposite"),
  seed = 1234,
  visit_effects_progressor = NULL,
  visit_effects_control = NULL,
  direction_by_cluster = NULL
)
}
\arguments{
\item{direction_by_cluster}{}
}
\description{
Title
}
