% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longAnalyses_plusModelSelection.R
\name{longAnalyses_plusModelSelection_scLASER}
\alias{longAnalyses_plusModelSelection_scLASER}
\title{Title}
\usage{
longAnalyses_plusModelSelection_scLASER(
  object,
  disease_var,
  time_var,
  covariates = NULL,
  subject_var = "subject_id",
  sample_var = "sample_id"
)
}
\arguments{
\item{object}{}

\item{disease_var}{}

\item{time_var}{}

\item{covariates}{}

\item{subject_var}{}

\item{sample_var}{}
}
\description{
Title
}
