% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLASER.R
\docType{class}
\name{scLASER-class}
\alias{scLASER-class}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{metadata}}{data.frame.}

\item{\code{pcs}}{matrixOrNULL.}

\item{\code{masc}}{dfOrNULL.}

\item{\code{umap}}{dfOrNULL.}

\item{\code{harmony}}{matrixOrNULL.}

\item{\code{nam_pcs}}{matrixOrNULL.}

\item{\code{pipeline_output}}{dfOrNULL.}
}}

