% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLASER_modeling.R
\name{scLASER_modeling}
\alias{scLASER_modeling}
\title{Title}
\usage{
scLASER_modeling(
  object,
  disease_var = "disease",
  time_var = "visit",
  covariates = c("age", "sex"),
  subject_var = "subject_id",
  sample_var = "sample_id"
)
}
\arguments{
\item{object}{}
}
\description{
Title
}
