% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLASER-show.R
\name{show,scLASER-method}
\alias{show,scLASER-method}
\title{Show method for scLASER objects}
\usage{
\S4method{show}{scLASER}(object)
}
\arguments{
\item{object}{A scLASER object.}
}
\description{
Custom display when a scLASER object is printed.
}
