% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_umap.R
\name{compute_umap}
\alias{compute_umap}
\title{Title}
\usage{
compute_umap(object, n_neighbors = 30, metric = "cosine", min_dist = 0.01, ...)
}
\arguments{
\item{object}{}

\item{n_neighbors}{}

\item{metric}{}

\item{min_dist}{}

\item{...}{}
}
\description{
Title
}
