% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_dummy_data_2time.R
\name{generate_dummy_data_2time}
\alias{generate_dummy_data_2time}
\title{Title}
\usage{
generate_dummy_data_2time(
  n_cells = 3000,
  sd_celltypes = 0.1,
  n_major_cell_types = 7,
  n_minor_cell_types = 3,
  relative_abundance = 0.1,
  n_major_interact_celltypes = 1,
  n_minor_interact_celltypes = 1,
  n_individuals = 30,
  n_batchs = 4,
  interaction_feature = c("visit"),
  time_points = 2,
  test_var = c("disease"),
  prop_disease = 0.5,
  fc_interact = 0.1,
  interaction_type = c("specific", "differential", "opposite"),
  seed = 1234
)
}
\arguments{
\item{n_cells}{}

\item{sd_celltypes}{}

\item{n_major_cell_types}{}

\item{n_minor_cell_types}{}

\item{relative_abundance}{}

\item{n_major_interact_celltypes}{}

\item{n_minor_interact_celltypes}{}

\item{n_individuals}{}

\item{n_batchs}{}

\item{interaction_feature}{}

\item{time_points}{}

\item{test_var}{}

\item{prop_disease}{}

\item{fc_interact}{}

\item{interaction_type}{}

\item{seed}{}
}
\description{
Title
}
