% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_pseudo_pcs_time.R
\name{generate_pseudo_pcs_time}
\alias{generate_pseudo_pcs_time}
\title{Title}
\usage{
generate_pseudo_pcs_time(
  data,
  n_pcs = 20,
  cluster_pcs = 1:20,
  disease_pcs = 0,
  sex_pcs = 0,
  age_pcs = 0,
  bmi_pcs = 0,
  batch_pcs = 0,
  interaction_pcs = 0,
  visit_pcs = 0,
  subject_pcs = 0,
  scale_factor = 2,
  cluster_ratio = 0.25,
  disease_ratio = 0,
  sex_ratio = 0,
  age_ratio = 0,
  bmi_ratio = 0,
  batch_ratio = 0,
  visit_ratio = 0,
  interaction_ratio = 0,
  subject_ratio = 0,
  cluster_col = "cell_type",
  disease_col = "disease",
  sex_col = "sex",
  age_col = "age",
  bmi_col = "bmi",
  batch_col = "batch",
  visit_col = "visit",
  interact_term_col = "interact_term",
  seed = 1234
)
}
\arguments{
\item{seed}{}
}
\description{
Title
}
