% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_harmony.R
\name{run_harmony}
\alias{run_harmony}
\title{Title}
\usage{
run_harmony(
  object,
  batch_col = "sample_id",
  theta = 2,
  do_pca = FALSE,
  epsilon.cluster = -Inf,
  epsilon.harmony = -Inf,
  max.iter.cluster = 30,
  max.iter.harmony = 10,
  plot_convergence = TRUE
)
}
\arguments{
\item{object}{}

\item{batch_col}{}

\item{theta}{}

\item{do_pca}{}

\item{epsilon.cluster}{}

\item{epsilon.harmony}{}

\item{max.iter.cluster}{}

\item{max.iter.harmony}{}

\item{plot_convergence}{}
}
\description{
Title
}
