% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLASER.R
\docType{class}
\name{scLASER-class}
\alias{scLASER-class}
\title{scLASER S4 class}
\description{
scLASER S4 class
}
\section{Slots}{

\describe{
\item{\code{metadata}}{data.frame}

\item{\code{pcs}}{matrix or NULL}

\item{\code{masc}}{data.frame or NULL}

\item{\code{umap}}{data.frame or NULL}

\item{\code{harmony}}{matrix or NULL}

\item{\code{nam_pcs}}{matrix or NULL}

\item{\code{NAM_matrix}}{matrix or NULL}

\item{\code{plsda_LV}}{matrix or NULL (latent variables from PLS-DA)}

\item{\code{pipeline_output}}{data.frame or NULL}
}}

