% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLASER.R
\name{scLASER}
\alias{scLASER}
\title{Create a scLASER object}
\usage{
scLASER(
  metadata = data.frame(),
  pcs = NULL,
  masc = NULL,
  umap = NULL,
  harmony = NULL,
  nam_pcs = NULL,
  NAM_matrix = NULL,
  plsda_LV = NULL,
  pipeline_output = NULL
)
}
\description{
Create a scLASER object
}
